refactor(cli): add shared cli-trace module and unify trace header keys
Move CLI_USER_AGENT and cliTraceHeaders to src/shared/cli-trace.ts so core,
services, and webhook code depend on a neutral layer instead of core importing
services. Use lowercase `user-agent` and `x-request-id` keys everywhere to
match ensureRequiredRequestHeaders in core/api.ts. Relocate unit tests
alongside the shared module.
