+++ src/services/extension/upload.ts
@@ -100,8 +100,8 @@ export function uploadArtifactEffect(
 
     for (let attempt = 1; attempt <= maxRetries; attempt++) {
       // Presigned S3 PUT: send only the headers returned with the URL (they are
-      // part of the AWS Signature v4 signing string). Do not add User-Agent or
-      // X-Request-ID here — unsigned headers can still affect behavior, and
+      // part of the AWS Signature v4 signing string). Do not add `user-agent` or
+      // `x-request-id` here — unsigned headers can still affect behavior, and
       // anything not in the signature can break the upload. Traceability for
       // this hop is the presigned URL request (GraphQL), not the PUT itself.
       const { "x-amz-meta-upload-id": _, ...headers } = target.requiredHeaders;
