+++ tests/unit/services/extension/presigned-url.test.ts
@@ -26,8 +26,8 @@ vi.mock("@/services/http-helpers", () => {
   return {
     getRequestHeaders: (token: string) => ({
       Authorization: `Bearer ${token}`,
-      "User-Agent": "godaddy-cli/0.0.0-test",
-      "X-Request-ID": "test-uuid",
+      "user-agent": "godaddy-cli/0.0.0-test",
+      "x-request-id": "test-uuid",
     }),
     makeGraphQLClientEffect: () =>
       Effect.succeed({
